Resident onboarding: optional app login, model/SQL updates
- Add CreateAppLogin and ApartmentAccessRoleCode to owner, co-owner, and tenant DTOs for optional user creation.
- Update creation flows to only require email if app login is requested; link Person to User/ApartmentUser as needed.
- Require IPasswordHasher in co-owner and tenant services.
- Switch ownership transfer/history IDs to long; add EffectiveDate.
- Support multiple owner assignments per unit with validation.
- Add/align SQL migration scripts for Persons, UnitOwners, Vehicles, OwnershipHistory, and Documents.
- Simplify controller exception handling.
- Update deed upload to use long IDs.
- Minor DTO/model doc and field alignment.
- Add two PDF files for testing/reference.

+++ Apartment_API/Apartment_API/Services/Interfaces/IOwnershipTransferResidentService.cs
@@ -9,5 +9,5 @@ public interface IOwnershipTransferResidentService
     Task<IReadOnlyList<OwnershipHistoryItemDto>> GetHistoryForUnitAsync(
         int apartmentId, int unitId, CancellationToken cancellationToken = default);
     Task<IdProofResultDto> UploadDeedAsync(
-        int apartmentId, int userId, int idOwnershipHistory, Stream fileStream, string fileName, CancellationToken cancellationToken = default);
+        int apartmentId, int userId, long idOwnershipHistory, Stream fileStream, string fileName, CancellationToken cancellationToken = default);
 }
